v1.0.4.9
- Added Pathing/Movement to Squad for when moving group of characters as a single unit
- Updated Animator to accept generic Something instead of Character to so can be used by both Character and Squad classes
- Added Flush option to KeyboardHandler
- Implemented wavering in Particle.Update()
- Updated ParticleManager for particle wavering
- Fixed World in Scene getting drawn twice when using default rendering
- Fixed SoundPaused/MusicPaused/AmbientPaused not used in SoundManager
- Replaced IFs in TimeHandler with Switches
- Fixed some bugs in TimeManager and added WeatherOptions to control which types of weather can be transitioned to
- Changed ALocation.Priority from int to float
- Replaced IFs in Weather.Update() with Switches
- Fixed WeatherType.Clear not being in the list of WeatherManager.Weathers which was preventing transition back to Clear weather in WeatherManager.ChangeWeather()
- Updated weather effects for Snow and Storm weather
- General code refactoring and bug fixing

+++ OP_Engine/Utility/ALocation.cs
@@ -8,7 +8,7 @@ namespace OP_Engine.Utility
 
         public int X;
         public int Y;
-        public int Priority;
+        public float Priority;
         public int Distance_ToStart;
         public int Distance_ToDestination;
         public ALocation Parent;
